Set up fragment that shows more information about the book the user's friend read when clicking on the cover in the social feed. Updated book and SocialFeed data class rating to a float. Storing more book information under the recent section on firebase.

+++ app/src/main/java/com/example/cranny/Book.kt
@@ -9,7 +9,7 @@ data class Book(
     var title: String,
     var authorNames: String?,
     var publicationDate: String?,
-    var starRating: Int?,
+    var starRating: Float?,
     var publisher: String?,
     var description: String?,
     var pageCount: Int?,
@@ -27,6 +27,4 @@ data class Book(
     var prevReadCount: Int?,
     var startDate: String,
     var endDate: String
-
-
 )
